requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
SQLAlchemy==2.0.27
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.6.1
python-multipart==0.0.9 